user sign-up and dosage calculator

+++ src/main/java/com/pooltracker/controllers/ClientController.java
@@ -5,6 +5,7 @@ import com.pooltracker.models.*;
 import com.pooltracker.models.data.AddressDao;
 import com.pooltracker.models.data.ClientDao;
 import com.pooltracker.models.data.PoolDao;
+import com.pooltracker.models.data.UserDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -30,11 +31,21 @@ public class ClientController {
     @Autowired
     private PoolDao poolDao;
 
+    @Autowired
+    private UserDao userDao;
+
     // Request path: /clients
     @RequestMapping(value = "")
-    public String index(Model model) {
+    public String index(Model model, @CookieValue(value = "user", defaultValue = "none") String email) {
 
-        model.addAttribute("clients", clientDao.findAll());
+        if(email.equals("none")) {
+            return "redirect:/login";
+        }
+
+        User user = userDao.findByEmail(email).get(0);
+
+        model.addAttribute("user", user.getBusiness());
+        model.addAttribute("clients", user.getClients());
         model.addAttribute("title", "My Clients");
 
         return "clients/index";
@@ -42,7 +53,12 @@ public class ClientController {
 
     // Adds client and associated classes. Displays add form.
     @RequestMapping(value = "add", method = RequestMethod.GET)
-    public String displayAddClientForm(Model model) {
+    public String displayAddClientForm(Model model, @CookieValue(value = "user", defaultValue = "none") String email) {
+
+        if(email.equals("none")) {
+            return "redirect:/login";
+        }
+        User user = userDao.findByEmail(email).get(0);
 
         model.addAttribute("title", "Add Client");
         model.addAttribute(new Client());
@@ -56,8 +72,13 @@ public class ClientController {
     //Add feature. checks validation and persists client.
     @RequestMapping(value = "add", method = RequestMethod.POST)
     public String processAddClientForm(@ModelAttribute @Valid Client newClient,
-                                       Errors errors, Model model) {
+                                       Errors errors, Model model,
+                                       @CookieValue(value = "user", defaultValue = "none") String email) {
 
+        if(email.equals("none")) {
+            return "redirect:/login";
+        }
+        User user = userDao.findByEmail(email).get(0);
 
         if (errors.hasErrors()) {
             model.addAttribute("title", "Add Client");
@@ -69,14 +90,23 @@ public class ClientController {
             return "clients/add";
         }
 
+        //adds user reference to client
+        newClient.setUser(user);
 
         clientDao.save(newClient);
         return "redirect:";
+
     }
 
     //Remove feature. displays all clients that are persisted.
     @RequestMapping(value = "remove", method = RequestMethod.GET)
-    public String displayRemoveClientForm(Model model) {
+    public String displayRemoveClientForm(Model model, @CookieValue(value = "user", defaultValue = "none") String email) {
+
+        if(email.equals("none")) {
+            return "redirect:/login";
+        }
+        User user = userDao.findByEmail(email).get(0);
+
         model.addAttribute("clients", clientDao.findAll());
         model.addAttribute("title", "Remove Client");
         return "clients/remove";
@@ -84,7 +114,13 @@ public class ClientController {
 
     //Remove feature. Allows a single deletion or multiple.
     @RequestMapping(value = "remove", method = RequestMethod.POST)
-    public String processRemoveClientForm(@RequestParam int[] clientIds) {
+    public String processRemoveClientForm(@RequestParam int[] clientIds, @CookieValue(value = "user", defaultValue = "none") String email) {
+
+        if(email.equals("none")) {
+            return "redirect:/login";
+        }
+        User user = userDao.findByEmail(email).get(0);
+
 
         for (int clientId : clientIds) {
             clientDao.delete(clientId);
@@ -96,38 +132,46 @@ public class ClientController {
 
     // Does NOT function at this time. need to work out update to the database.
     @RequestMapping(value = "edit/{clientId}")
-    public String displayEditForm(Model model, @PathVariable int clientId, State state) {
+    public String displayEditForm(Model model, @PathVariable int clientId, State state, FilterType filterType,
+                                  PumpHp pumpHp, PumpPh pumpPh,
+                                  @CookieValue(value = "user", defaultValue = "none") String email) {
+
+
+        if(email.equals("none")) {
+            return "redirect:/login";
+        }
+        User user = userDao.findByEmail(email).get(0);
+
         model.addAttribute("title", "Edit Client");
         Client c = clientDao.findOne(clientId);
-        Address a = addressDao.findOne(clientId);
         model.addAttribute("client", c);
         model.addAttribute("states", state.values());
-        model.addAttribute("filterTypes", FilterType.values());
-        model.addAttribute("pumpHps", PumpHp.values());
-        model.addAttribute("pumpPhs", PumpPh.values());
+        model.addAttribute("filterTypes", filterType.values());
+        model.addAttribute("pumpHps", pumpHp.values());
+        model.addAttribute("pumpPhs", pumpPh.values());
 
         return "clients/edit";
 
     }
+
     // Does NOT function at this time. need to work out update to the database.
     @RequestMapping(value = "edit/{clientId}", method = RequestMethod.POST)
-    public String processEditForm( @PathVariable int clientId, State state,
-                                   @ModelAttribute @Valid Client editClient,
-                                   Errors errors, Model model) {
+    public String processEditForm( @PathVariable int clientId, State state, FilterType filterType, PumpHp pumpHp,
+                                   PumpPh pumpPh, @ModelAttribute @Valid Client editClient,
+                                   Errors errors, Model model,
+                                   @CookieValue(value = "user", defaultValue = "none") String email) {
 
         if (errors.hasErrors()) {
             model.addAttribute("title", "Edit Client");
             model.addAttribute("client", editClient);
             model.addAttribute("state", state.values());
-            model.addAttribute("filterTypes", FilterType.values());
-            model.addAttribute("pumpHps", PumpHp.values());
-            model.addAttribute("pumpPhs", PumpPh.values());
+            model.addAttribute("filterTypes", filterType.values());
+            model.addAttribute("pumpHps", pumpHp.values());
+            model.addAttribute("pumpPhs", pumpPh.values());
             return "clients/edit";
         }
 
         Client updateClient = clientDao.findOne(clientId);
-        updateClient.setFirstName(editClient.getFirstName());
-        updateClient.setLastName(editClient.getLastName());
         clientDao.save(updateClient);
 
         return "redirect:";
@@ -136,7 +180,8 @@ public class ClientController {
 
     //Displays the personal info of a added client by its id.
     @RequestMapping(value = "client-info/{id}", method = RequestMethod.GET)
-    public String displayClientInfo(Model model, @PathVariable int id) {
+    public String displayClientInfo(Model model, @PathVariable int id,
+                                    @CookieValue(value = "user", defaultValue = "none") String email) {
 
         model.addAttribute("title", "Client Information");
         Client c = clientDao.findOne(id);
@@ -151,7 +196,8 @@ public class ClientController {
 
     //Displays the pool info of a added client by its id.
     @RequestMapping(value = "pool-info/{id}", method = RequestMethod.GET)
-    public String displayPoolInfo(Model model, @PathVariable int id) {
+    public String displayPoolInfo(Model model, @PathVariable int id,
+                                  @CookieValue(value = "user", defaultValue = "none") String email) {
 
         model.addAttribute("title", "Pool Information");
         Client c = clientDao.findOne(id);
@@ -166,14 +212,13 @@ public class ClientController {
 
     //Displays the dosage view of a added client by its id.
     @RequestMapping(value = "dosage-view/{id}", method = RequestMethod.GET)
-    public String displayDosageInfo(Model model, @PathVariable int id) {
+    public String displayDosageInfo(Model model, @PathVariable int id,
+                                    @CookieValue(value = "user", defaultValue = "none") String email) {
 
         Client c = clientDao.findOne(id);
         Pool p = poolDao.findOne(id);
         float gallons = p.getGallons();
         Dosage d = new Dosage(gallons);
-
-        //System.out.println(d.toString());
 
         model.addAttribute("title", "Dosage Information");
         model.addAttribute("client", c);
